08_openai-agent-sdk/examples/expense_tracker_agent/pyproject.toml
[project]
name = "expense-tracker-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.11",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]
